cleaned up a bit and error container handling
