+++ srcs/containers/django/src/pong/static/js/loadPage.js
@@ -1,4 +1,4 @@
-import { handleFormSubmitWrapper, handleLogout } from './userMgmt.js';
+import { handleFormSubmitWrapper, handleLogout, fetchUserInfo } from './userMgmt.js';
 
 function loadPageClosure(game) {
 	return async (page) => {
@@ -37,55 +37,6 @@ function loadPageClosure(game) {
 			throw error;
 		}
 	};
-}
-
-// async function fetchUserInfo() {
-// 	try {
-// 	  const response = await fetch('/api/home/');
-// 	  const data = await response.json();
-// 	  if (data.user_info.is_logged_in) {
-// 		window.userInfo = data; // Store userInfo in a global variable
-// 		console.log('User info fetched:', window.userInfo);
-// 	  } else {
-// 		window.userInfo = null; // Clear userInfo if not logged in
-// 	  }
-// 	} catch (error) {
-// 	  console.error('Failed to fetch user info', error);
-// 	  window.userInfo = null; // Ensure userInfo is null on failure
-// 	}
-//   }
-
-// async function updateUI() {
-// 	await fetchUserInfo();
-// 	// console.log('User info:', window.userInfo.user_info.username);
-// 	const userStatus = window.userInfo;
-//     if (userStatus && userStatus.is_logged_in) {
-// 		console.log('User info:', userStatus.user_info.username);
-//         document.getElementById('login-link').style.display = 'none';
-//         document.getElementById('register-link').style.display = 'none';
-//         document.getElementById('logout-link').style.display = 'block';
-//         document.getElementById('user-info').style.display = 'block';
-//         document.getElementById('user-info').innerText = `Welcome, ${userStatus.user_info.username}`;
-//     } else {
-// 		document.getElementById('login-link').style.display = 'block';
-//         document.getElementById('register-link').style.display = 'block';
-//         document.getElementById('logout-link').style.display = 'none';
-//         document.getElementById('user-info').style.display = 'none';
-//     }
-// }
-
-async function fetchUserInfo() {
-    try {
-        const response = await fetch('/api/userinfo/');
-        // if (!response.ok) {
-        //     throw new Error(`HTTP error! Status: ${response.status}`);
-        // }
-        const data = await response.json();
-        return data.user_info; // Return user_info object directly
-    } catch (error) {
-        // console.error('Failed to fetch user info', error);
-        return null; // Return null if there's an error
-    }
 }
 
 async function updateUI() {
