+++ srcs/containers/django/src/pong/templates/main/base.html
@@ -22,7 +22,7 @@
 
   <body>
   <div id="threejs-container"></div>
-  <div id="error-container" class="container-fluid" style="display: none;"></div>
+  <div id="error-container" class="container-fluid mx-auto p-4 fixed-top d-flex text-start" style="display: none;"></div>
   <div id="notification" class="alert alert-success alert-dismissible fade show" role="alert" style="display: none;">
     <strong>Holy guacamole!</strong> You should check in on some of those fields below.
     <button type="button" class="close" data-dismiss="alert" aria-label="Close">
